+++ tests/testthat/test-vec-utils.R
@@ -6,11 +6,6 @@ test_that("vector is modified", {
   expect_equal(out, list(1, b = 20, c = "30", 4, 5, 6))
 })
 
-test_that("are_na() requires vector input but not is_na()", {
-  expect_error(are_na(base::eval), "must be a vector")
-  expect_false(is_na(base::eval))
-})
-
 test_that("seq2() creates increasing sequences", {
   expect_identical(seq2(2, 3), 2:3)
   expect_identical(seq2(3, 2), int())